requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlmodel>=0.0.8
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
slowapi>=0.1.8
pytest>=7.4.0
httpx>=0.24.0
redis>=5.0.0
fakeredis>=2.20.0
boto3>=1.34.0
aiofiles>=23.2.1 